more edits, added redshift

+++ anaconda/anaconda.sh
@@ -10,16 +10,22 @@ dir=$root/anaconda
 # mkdir anaconda
 # cd anaconda
 
-# dependencies
-apt-get install libgl1-mesa-glx libegl1-mesa libxrandr2 libxrandr2 libxss1 libxcursor1 libxcomposite1 libasound2 libxi6 libxtst6
+# checks if anaconda already installed
+if ! test -d ~/anaconda3
+then
+  # dependencies
+  apt-get install libgl1-mesa-glx libegl1-mesa libxrandr2 libxrandr2 libxss1 libxcursor1 libxcomposite1 libasound2 libxi6 libxtst6
 
-wget https://repo.anaconda.com/archive/Anaconda3-2019.07-Linux-x86_64.sh
-chmod u+x Anaconda3-2019.07-Linux-x86_64.sh
+  wget https://repo.anaconda.com/archive/Anaconda3-2019.07-Linux-x86_64.sh
+  chmod u+x Anaconda3-2019.07-Linux-x86_64.sh
 
-# runs anaconda setup script (IN BASH)
-bash ./Anaconda3-2019.07-Linux-x86_64.sh -y # requires you to click thru
+  # runs anaconda setup script (IN BASH)
+  bash ./Anaconda3-2019.07-Linux-x86_64.sh -y # requires you to click thru
 
-rm ./Anaconda3-2019.07-Linux-x86_64.sh
+  rm ./Anaconda3-2019.07-Linux-x86_64.sh
+else
+  printf "\nSUCCESS - CONDA ALREADY INSTALLED\n"
+fi
 
 # updates conda if already installed
 conda update -n base -c defaults conda -y
@@ -27,13 +33,13 @@ conda update -n base -c defaults conda -y
 # instantiates conda venv's in the shell
 printf "\nCALLS CONDA INIT\n"
 # conda init, then rerun bashrc
-conda init
+conda init bash
 source ~/.bashrc
 
 # calls files to create custom envs
 # async errors for `sh` vs `source`?
-source $dir/dev.sh
-source $dir/nondev.sh
+bash $dir/dev.sh
+bash $dir/nondev.sh
 
 # test
 conda env list
